fix(worker): version-specific request URLs, delivered bridge inputs, real ownership gate (R9R-001, R9R-002, R9R-003, R9R-005)
R9R-001: the supervisor-owned Worker resolves a VERSION-SPECIFIC download path
for each update request, so the artifact coordinates it records name the
sandbox's pinned version (fail-closed server-side) rather than the global pin.

R9R-002: `maybe_bridge_to_supervisor` derives bridge inputs from the
server-delivered heartbeat inputs when the on-disk config carries none (the
legacy-migration path), materializing BOTH the Supervisor config and a
supervisor-owned Worker config before spawning.

R9R-003: gate bridge completion on ACTUAL ownership evidence — Supervisor live
AND AnyHarness /health passing — not just a Supervisor PID. The Worker child is
deliberately not part of the in-process gate (the exclusive worker flock makes
pre-handoff child enrollment a deadlock); documented inline.

R9R-005: split the 1167-line supervisor_bridge.rs into
supervisor_bridge/{mod,mailbox,bridge/mod,bridge/tests}.rs, all under 600.
